MARBLE-1544 Fix problem with manifests (#296)
* We have a problem where manifests have thumbnail.id and service.id pointing to collection/filepath instead of just filepath.

This problem arose because we have changed image processing to be based on AppSync instead of standard.json, but we still use the standard.json to create manifests.  And since we have changed the location of images for image processing to accommodate instances where one image is used for multiple items by changing the image id (and path).

It was also pointed out that the old url was wrong, and needed "/" replaced with "%2F" in the imageId portion of the path.

To Fix this, I updated csv_collection.py _file_paths.  This defines iiifUri and iiifImageUri, which the manifest process later uses.

* Removed some commented-out code from csv_collection.py
* Added some ArchviesSpace Urls to harvest.