MARBLE-2105 Update Regex pattern for Pfefer Wrestling Collection
The previous regular expression worked with 3 exceptions:

* PFE_700-1-12_001r
* PFE_700-1-15_001r
* PFE_700-01-60A_0001r

The first two appear to be inconsistently numbered `1` instead of `01` and I will be asking that these two series are renamed to confrom to the pattern.
The third is a different case where there should be a `60` and a `60A` due to a historical mis-numbering situation. The PR addresses this specific case (and other similar potential cases), by adding in an option single alphabetical character after the second `[0-9]{2}`.